New commands make chart.json, budget.json, initial.json
New commands parse Church Windows output to construct the chart of
accounts, the budget, and the initial account balances.  This needs to
be done only once at the start of the year.  Extracting this from
report generation means the only additional data report generation
needs is the Church Windows dump of the transaction log.